submission/solution/requirements.txt
numpy
opencv-python
pillow
torch
